Deleting unnecessary pages
