+++ doce_conta/lib/view/Initial/successful_register.dart
@@ -1,4 +1,4 @@
-import 'package:doce_conta/view/login_screen.dart';
+import 'package:doce_conta/view/Initial/login_screen.dart';
 import 'package:flutter/material.dart';
 
 class SuccessfulRegister extends StatefulWidget {
