+++ doce_conta/lib/view/Product/hub_products.dart
@@ -1,5 +1,4 @@
-import 'package:doce_conta/view/product_screen.dart';
-import 'package:doce_conta/view/product_selection_screen.dart';
+import 'package:doce_conta/view/Product/product_selection_screen.dart';
 import 'package:doce_conta/widgets/container_custom_button.dart';
 import 'package:flutter/material.dart';
 
@@ -22,7 +21,7 @@ class _HubProducts extends State<HubProducts> {
       appBar: AppBar(
         backgroundColor: const Color(0xff003326),
         title: Text(
-          "O que deseja Precificar?",
+          "Produtos para precificação",
           style: TextStyle(color: Colors.white),
         ),
         centerTitle: true,
