+++ doce_conta/lib/view/Reports/annual_report.dart
@@ -1,8 +1,7 @@
-import 'package:doce_conta/view/opening_for_monthly_report.dart';
+import 'package:doce_conta/view/Reports/opening_for_monthly_report.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
-import '../widgets/appBarWithIcon.dart';
-import '../widgets/conteiner_text_center.dart';
+import '../../widgets/conteiner_text_center.dart';
 
 void main(){
   runApp(const AnnualReportPage());
