+++ doce_conta/lib/view/Reports/monthly_report_screen.dart
@@ -1,22 +1,22 @@
-import 'package:doce_conta/view/cost_item_addition.dart';
+import 'package:doce_conta/view/Reports/report_descriptions_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
-class InfraestructureCost extends StatelessWidget {
+class MontlhyReport extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: Colors.white,
+      backgroundColor: Color(0xFF98CFC2),
       appBar: AppBar(
         backgroundColor: const Color(0xFF00382B),
         leading: IconButton(
           icon: const Icon(Icons.arrow_back, color: Colors.white),
           onPressed: () {
-            // Navigator.of(context).pop();
+            Navigator.of(context).pop();
           },
         ),
         title: SvgPicture.asset(
-          'assets/images/infraestrutura.svg',
+          'assets/images/relatorio.svg',
           height: 40,
           color: const Color(0xFFF5A001),
         ),
@@ -28,12 +28,12 @@ class InfraestructureCost extends StatelessWidget {
           Expanded(
             child: ListView.separated(
               padding: const EdgeInsets.all(16.0),
-              itemCount: 4,
+              itemCount: 5,
               itemBuilder: (context, index) {
                 // Lista de produtos
-                final serviceNames = ['Aluguel', 'Luz', 'Água', 'Segurança'];
-                final servicePrice = ['R\$ 1400', 'R\$ 350', 'R\$ 180', 'R\$ 380'];
-                return _buildServiceItem(serviceNames[index], servicePrice[index]);
+                final infoName = ['Quantidade de itens vendidos', 'Valor bruto: caixa', 'Lucro diário', 'Valor Contas', 'Reserva'];
+                final infoPrice = ['x', 'R\$', 'R\$', 'R\$', 'R\$'];
+                return _buildInfoItem(infoName[index], infoPrice[index]);
               },
               separatorBuilder: (context, index) => Divider(
                 color: Colors.amber, // Linha amarela entre os itens
@@ -45,16 +45,19 @@ class InfraestructureCost extends StatelessWidget {
           // Botão de adicionar
           Padding(
             padding: const EdgeInsets.only(bottom: 16.0),
-            child: FloatingActionButton(
+            child: ElevatedButton(
               onPressed: () {
-                // Adicionar novo produto
+                //Linkar página de descrições
                 Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                        builder: (context) => CostItemAddition()));
+                  context,
+                  MaterialPageRoute(
+                    builder: (context) => ReportDescriptions()));
               },
-              child: const Icon(Icons.add, color: Colors.black),
-              backgroundColor: Colors.white,
+              child: 
+              const Text(
+                "Descrições",
+                style: TextStyle(color: Colors.black),
+              )
             ),
           ),
         ],
@@ -62,7 +65,7 @@ class InfraestructureCost extends StatelessWidget {
     );
   }
 
- Widget _buildServiceItem(String serviceName, String servicePrice) {
+ Widget _buildInfoItem(String infoName, String infoPrice) {
     return Container(
       margin: const EdgeInsets.symmetric(vertical: 8.0),
       padding: const EdgeInsets.symmetric(horizontal: 16.0, vertical: 12.0),
@@ -75,11 +78,11 @@ class InfraestructureCost extends StatelessWidget {
         mainAxisAlignment: MainAxisAlignment.spaceBetween,
         children: [
           Text(
-            serviceName,
+            infoName,
             style: const TextStyle(fontSize: 16, color: Colors.black),
           ),
           Text(
-            servicePrice,
+            infoPrice,
             style: const TextStyle(fontSize: 16, color: Colors.black),
           ),
         ]
